access to submodules without separate import statements
