+++ rosys/test/log_configuration.py
@@ -5,6 +5,7 @@ import sys
 
 import coloredlogs
 import icecream
+
 import rosys
